make.py: Move boards definition to boards.py.
